requirements.txt
langchain-groq>=0.0.3
langchain-core>=0.1.0
langchain-community>=0.0.10
pypdf>=3.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
requests>=2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
rouge-score>=0.1.2
bert-score>=0.3.11
pandas>=1.5.0
rouge>=1.0.1 